Add awesomebot link checker GitHub Action
awesomebot will check all the links in a markdown file.

This configuration is modified from the one I use to check the
[awesome-zsh-plugins](https://github.com/unixorn/awesome-zsh-plugins)
list.

It's configured to ignore 5xx return codes to help prevent false
positives for broken links and to wait 1 second between probes so that
it doesn't get rate-limited by GitHub as it checks links there.

Signed-off-by: Joe Block <jpb@unixorn.net>